test-requirements.txt
ansible-lint==5.4.0
molecule==3.4.0
molecule-docker==0.2.4
docker==5.0.0
yamllint
flake8
pytest
pytest-testinfra
tox
tox-gh-actions
netaddr==0.8.0
jmespath
